backend/requirements.txt
fastapi
uvicorn
python-socketio
python-multipart
boto3
python-dotenv